+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [#90](https://github.com/simonneutert/papierkram_api_client/pull/90) crud ops for time_entries. [@simonneutert](https://github.com/simonneutert)
 - [#89](https://github.com/simonneutert/papierkram_api_client/pull/89) CRUD for tracker tasks. [@simonneutert](https://github.com/simonneutert)
+- [#88](https://github.com/simonneutert/papierkram_api_client/pull/88) Adds Income::PaymentTerm endpoints. [@simonneutert](https://github.com/simonneutert)
 
 ## [0.3.1] - 2023-10-05